fix missing configuration
To populate $conf['plugin']['nsexport'] we need to call $this->getConf()
at least once before.

Also it turns out that configuration keys with 4 underscrores are split
into a multi-level array. This is undocumented and possible buggy
behavior for plugins, though it is intended for core. Effectively this
meant that changes to packer____ziphtml____zip and
packer____ziphtml____compress were ignored.

The workaround is simply to use only 1 underscore in the plugin name.